pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "probeflow"
version = "0.0.0b0"  # beta — surfaces and on-disk formats may change
description = "ProbeFlow — STM scan browser, processor, and Createc↔Nanonis toolkit. Browse .sxm scans with colormaps, apply line corrections / background subtraction / smoothing / FFT / grain detection, and convert Createc .dat files to Nanonis .sxm + PNG from either a GUI or a bash-friendly CLI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "scipy",
  "pillow",
  "PySide6",
  "matplotlib",
]

[project.optional-dependencies]
dev      = ["pytest"]
features = ["opencv-python", "scikit-learn"]
gwyddion = ["gwyfile"]
all      = ["opencv-python", "scikit-learn", "gwyfile", "pytest"]

[project.scripts]
probeflow = "probeflow.cli:main"
dat-sxm   = "probeflow.io.converters.createc_dat_to_sxm:main"
dat-png   = "probeflow.io.converters.createc_dat_to_png:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["probeflow", "probeflow.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
